requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.4.0
torchvision==0.5.0
